requirements.txt
pytorch-lightning==1.5.10
torch==1.11.0
requests==2.27.1
torchvision==0.12.0
flask==2.0.2
pillow==8.2.0
numpy==1.21.3
onnxruntime==1.10.0